Add files via upload
House prices are a crucial indicator of the economic condition, and it is also a point of great interest for buyers and sellers.  Both parties are interested in the factors that affect the price. From one side, the buyers are interested in evaluating the house price to understand what they can afford and what factors affect the price. The correct understanding of house price drivers will help a buyer find the house with suitable parameters according to their budget. At the same time, it is evident that buyers do not want to overpay. So, they need to know the house's fair price. On the seller side (private or corporate), they need to estimate the asset's value they possess. It will help forecast potential future revenue, estimate the market conjecture (prices, competitors, consumers), elaborate the efficient promotional campaign, etc. 

Our project's primary purpose is to implement the knowledge, instruments, and techniques learned in class in solving real-world problems. In particular, we want to build a linear regression model that gives accurate predictions of the house prices based on the dataset. Additionally, to see if there is any correlation based on the variables and how they play in the sales price.

The dataset we will use for this project is the "Ames Housing Data" created by Dean De Cock. As a part of his work as a professor, he used the famous Boston Housing Data Set for his regression class (Carnegie Mellon University, 1980). However, this dataset only contains 506 observations and 14 variables and is quite outdated: the original dataset came from the 70s, resulting in data not corresponding to today's modern real estate market. To obtain a more relevant dataset, Dean De Cock, in collaboration with the students of Iowa State StatCom and the Ames City Assessor's Office, created the new Ames Housing data set (Dean, 2011). The data describes the sale of individual residential property in Ames, Iowa, from 2006 to 2010. The original dataset contains 2930 observations and many explanatory variables (23 nominal, 23 ordinal, 14 discrete, and 20 continuous) with some missing values. Involving in assessing home values in such as lot size in square feet, number of fireplaces, the pool area in square feet, number of kitchens, kitchen quality, type of roof, overall material, finish quality, overall condition rating, the property's sale price in dollars, the building class, the general zoning classification, proximity to the main road or railroad, type of dwelling, year build, etc. The dataset available at Kaggle contains 2920 observations equally divided into the training and test set where the test does not include the values of target variables. More information can be found at the Kaggle or in the article that describes this dataset.